+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 
 0.1 (unreleased)
 ----------------
+- Added portlet that shows Collection widget defined on a faceted nav enabled folder.
+  [gbastien]
 
 - Initial release.
   [IMIO]
-